requirements_mlx.txt
mlx
mlx-lm
numpy
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub